pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "weekly-product-review-pulse"
version = "0.1.0"
description = "AI agent scaffold for a weekly product review pulse."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0",
  "fastembed>=0.8.0,<1.0",
  "google-play-scraper>=1.2.7,<2.0",
  "hdbscan>=0.8.40,<1.0",
  "httpx>=0.27,<1.0",
  "jinja2>=3.1,<4.0",
  "numpy>=1.26,<3.0",
  "openai>=1.30,<2.0",
  "opentelemetry-api>=1.28,<2.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28,<2.0",
  "opentelemetry-sdk>=1.28,<2.0",
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "PyYAML>=6.0,<7.0",
  "scikit-learn>=1.5,<2.0",
  "sqlite-vec>=0.1.6,<0.2.0",
  "structlog>=24.1,<25.0",
  "tzdata>=2024.1,<2027.0",
  "typer>=0.12,<1.0",
  "umap-learn>=0.5.7,<1.0",
  "uvicorn>=0.30.0,<1.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10,<2.0",
  "pytest>=8.2,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5.0,<0.6.0",
  "types-PyYAML>=6.0.12,<7.0.0",
]

[project.scripts]
pulse = "agent.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "SIM", "UP"]

[tool.mypy]
python_version = "3.12"
packages = ["agent"]
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
  "fastapi",
  "fastapi.*",
  "fastembed",
  "google_play_scraper",
  "hdbscan",
  "opentelemetry",
  "opentelemetry.*",
  "sklearn",
  "sklearn.*",
  "umap",
  "umap.*",
  "uvicorn",
]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["agent"]
